+++ qiskit/providers/ibmq/api_v2/rest/job.py
@@ -23,10 +23,14 @@ class Job(RestAdapterBase):
     """Rest adapter for job related endpoints."""
 
     URL_MAP = {
+        'callback_upload': '/jobDataUploaded',
         'cancel': 'cancel',
+        'download_url': '/jobDownloadUrl',
         'self': '',
         'status': '/status',
-        'properties': '/properties'
+        'properties': '/properties',
+        'result_url': '/resultDownloadUrl',
+        'upload_url': '/jobUploadUrl'
     }
 
     def __init__(self, session, job_id):
@@ -62,14 +66,29 @@ class Job(RestAdapterBase):
 
         return response
 
+    def callback_upload(self):
+        """Notify the API after uploading a Qobj via object storage."""
+        url = self.get_url('callback_upload')
+        return self.session.post(url).json()
+
     def cancel(self):
         """Cancel a job."""
         url = self.get_url('cancel')
         return self.session.post(url).json()
 
+    def download_url(self):
+        """Return an object storage URL for downloading the Qobj."""
+        url = self.get_url('download_url')
+        return self.session.get(url).json()
+
     def properties(self):
         """Return the backend properties of a job."""
         url = self.get_url('properties')
+        return self.session.get(url).json()
+
+    def result_url(self):
+        """Return an object storage URL for downloading results."""
+        url = self.get_url('result_url')
         return self.session.get(url).json()
 
     def status(self):
@@ -77,9 +96,42 @@ class Job(RestAdapterBase):
         url = self.get_url('status')
         return self.session.get(url).json()
 
+    def upload_url(self):
+        """Return an object storage URL for uploading the Qobj."""
+        url = self.get_url('upload_url')
+        return self.session.get(url).json()
+
+    def put_object_storage(self, url, qobj_dict):
+        """Upload a Qobj via object storage.
+
+        Args:
+            url (str): object storage URL.
+            qobj_dict (dict): the qobj to be uploaded, in dict form.
+
+        Returns:
+            str: text response, that will be empty if the request was
+                successful.
+        """
+        response = self.session.put(url, json=qobj_dict, bare=True)
+        return response.text
+
+    def get_object_storage(self, url):
+        """Get via object_storage.
+
+        Args:
+            url (str): object storage URL.
+
+        Returns:
+            dict: json response.
+        """
+        return self.session.get(url, bare=True).json()
+
 
 def build_url_filter(excluded_fields, included_fields):
     """Return a URL filter based on included and excluded fields.
+
+    If a field appears in both excluded_fields and included_fields, it
+    is ultimately included.
 
     Args:
         excluded_fields (list[str]): names of the fields to explicitly
@@ -92,19 +144,21 @@ def build_url_filter(excluded_fields, included_fields):
     """
     excluded_fields = excluded_fields or []
     included_fields = included_fields or []
-    fields_bool = {}
+    field_flags = {}
     ret = {}
 
     # Build a map of fields to bool.
     for field_ in excluded_fields:
-        fields_bool[field_] = False
+        field_flags[field_] = False
     for field_ in included_fields:
-        fields_bool[field_] = True
+        # Set the included fields. If a field_ here was also in
+        # excluded_fields, it is overwritten here.
+        field_flags[field_] = True
 
-    if 'properties' in fields_bool:
-        fields_bool['calibration'] = fields_bool.pop('properties')
+    if 'properties' in field_flags:
+        field_flags['calibration'] = field_flags.pop('properties')
 
-    if fields_bool:
-        ret = {'fields': fields_bool}
+    if field_flags:
+        ret = {'fields': field_flags}
 
     return ret