+++ qiskit/providers/ibmq/ibmqbackend.py
@@ -25,9 +25,9 @@ from qiskit.providers.models import (BackendStatus, BackendProperties,
                                      PulseDefaults)
 
 from .api import ApiError
-from .api.apijobstatus import ApiJobStatus
+from .apiconstants import ApiJobStatus, ApiJobKind
 from .exceptions import IBMQBackendError, IBMQBackendValueError
-from .ibmqjob import IBMQJob
+from .job import IBMQJob
 
 logger = logging.getLogger(__name__)
 
@@ -54,7 +54,7 @@ class IBMQBackend(BaseBackend):
         self.project = credentials.project
 
     def run(self, qobj):
-        """Run qobj asynchronously.
+        """Run a Qobj asynchronously.
 
         Args:
             qobj (Qobj): description of job
@@ -181,7 +181,9 @@ class IBMQBackend(BaseBackend):
                                                   filter=api_filter)
         job_list = []
         for job_info in job_info_list:
-            if job_info.get('kind', None) != 'q-object':
+            try:
+                ApiJobKind(job_info.get('kind', None))
+            except ValueError:
                 # Discard pre-qobj jobs.
                 break
 
@@ -213,17 +215,20 @@ class IBMQBackend(BaseBackend):
                                        .format(job_id, job_info['error']))
 
             # Check for jobs from a different backend.
-            if job_info['backend']['name'] != self.name():
+            job_backend_name = job_info['backend']['name']
+            if job_backend_name != self.name():
                 warnings.warn('Job "{}" belongs to another backend than the one queried. '
                               'The query was made on backend "{}", '
                               'but the job actually belongs to backend "{}".'
-                              .format(job_id, job_info['backend']['name'], self.name()))
+                              .format(job_id, self.name(), job_backend_name))
                 raise IBMQBackendError('Failed to get job "{}": '
                                        'job does not belong to backend "{}".'
-                                       .format(job_id, job_info['backend']['name']))
+                                       .format(job_id, self.name()))
 
             # Check for pre-qobj jobs.
-            if job_info.get('kind', None) != 'q-object':
+            try:
+                ApiJobKind(job_info.get('kind', None))
+            except ValueError:
                 warnings.warn('The result of job {} is in a no longer supported format. '
                               'Please send the job using Qiskit 0.8+.'.format(job_id),
                               DeprecationWarning)