+++ qiskit/providers/ibmq/ibmqsingleprovider.py
@@ -25,6 +25,7 @@ from qiskit.validation.exceptions import ModelValidationError
 
 from .api import IBMQConnector
 from .api_v2 import IBMQClient
+from .api_v2.exceptions import AuthenticationLicenseError
 from .ibmqbackend import IBMQBackend, IBMQSimulator
 
 
@@ -49,6 +50,7 @@ class IBMQSingleProvider(BaseProvider):
 
         # Get a connection to IBMQ.
         self.credentials = credentials
+        self.is_new_api = False
         self._api = self._authenticate(self.credentials)
         self._ibm_provider = ibmq_provider
 
@@ -64,8 +66,7 @@ class IBMQSingleProvider(BaseProvider):
 
         return filter_backends(backends, filters=filters, **kwargs)
 
-    @classmethod
-    def _authenticate(cls, credentials):
+    def _authenticate(self, credentials):
         """Authenticate against the IBMQ API.
 
         Args:
@@ -76,29 +77,41 @@ class IBMQSingleProvider(BaseProvider):
         Raises:
             ConnectionError: if the authentication resulted in error.
         """
-        # TODO: add more robust detection.
-        # Check if the URL belongs to auth services of the new API.
-        if ('quantum-computing.ibm.com/api' in credentials.url and
-                'auth' in credentials.url):
-            return IBMQClient(api_token=credentials.token,
-                              auth_url=credentials.url)
+        # Use a temporary IBMQConnector for determining API version.
+        # TODO: replace with a IBMQClient or a Session directly after support
+        # for proxies is tested for RetrySession.
+        # Prepare the config_dict for IBMQConnector.
+        config_dict = {
+            'url': credentials.url,
+        }
+        if credentials.proxies:
+            config_dict['proxies'] = credentials.proxies
+        if credentials.websocket_url:
+            config_dict['websocket_url'] = credentials.websocket_url
 
+        # Prepare the version_config for the version detector.
+        version_config = config_dict.copy()
+        # By passing "access_token" in the dict, we bypass the login.
+        version_config['access_token'] = 'version_check'
+        version_connector = IBMQConnector(None, version_config,
+                                          credentials.verify)
+
+        # Check if the URL belongs to auth services of the new API.
         try:
-            config_dict = {
-                'url': credentials.url,
-            }
-            if credentials.proxies:
-                config_dict['proxies'] = credentials.proxies
-            if credentials.websocket_url:
-                config_dict['websocket_url'] = credentials.websocket_url
-            return IBMQConnector(credentials.token, config_dict,
-                                 credentials.verify)
+            version_info = version_connector.api_version()
+            self.is_new_api = version_info['new_api']
+
+            if version_info['new_api'] and 'api-auth' in version_info:
+                return IBMQClient(api_token=credentials.token,
+                                  auth_url=credentials.url)
+            else:
+                return IBMQConnector(credentials.token, config_dict,
+                                     credentials.verify)
+        except AuthenticationLicenseError as ex:
+            raise ConnectionError("Couldn't connect to IBMQ server: {0}"
+                                  .format(ex)) from None
         except Exception as ex:
             root_exception = ex
-            if 'License required' in str(ex):
-                # For the 401 License required exception from the API, be
-                # less verbose with the exceptions.
-                root_exception = None
             raise ConnectionError("Couldn't connect to IBMQ server: {0}"
                                   .format(ex)) from root_exception