+++ test/ibmq/test_ibmq_client_v2.py
@@ -15,7 +15,7 @@
 """Tests for the IBMQClient for API v2."""
 
 import re
-from unittest import SkipTest, skip
+from unittest import skip
 
 from qiskit.circuit import ClassicalRegister, QuantumCircuit, QuantumRegister
 from qiskit.compiler import assemble, transpile
@@ -23,6 +23,8 @@ from qiskit.providers.ibmq import IBMQ
 from qiskit.providers.ibmq.api_v2 import IBMQClient
 from qiskit.providers.ibmq.api_v2.exceptions import ApiError
 from qiskit.test import QiskitTestCase, requires_qe_access
+
+from ..decorators import requires_new_api_auth
 
 
 class TestIBMQClient(QiskitTestCase):
@@ -45,21 +47,19 @@ class TestIBMQClient(QiskitTestCase):
     @staticmethod
     def _get_client(qe_token, qe_url):
         """Helper for instantiating an IBMQClient."""
-        if not ('quantum-computing.ibm.com/api' in qe_url and
-                'auth' in qe_url):
-            raise SkipTest('Test needs credentials for new API')
         return IBMQClient(qe_token, qe_url)
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_valid_login(self, qe_token, qe_url):
         """Test valid authenticating against IBM Q."""
         client = self._get_client(qe_token, qe_url)
-        self.assertTrue(client.api_client.session.access_token)
+        self.assertTrue(client.client_api.session.access_token)
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_run_job(self, qe_token, qe_url):
         """Test running a job against a simulator."""
-        _ = self._get_client(qe_token, qe_url)
         IBMQ.enable_account(qe_token, qe_url)
 
         # Create a Qobj.
@@ -70,12 +70,42 @@ class TestIBMQClient(QiskitTestCase):
 
         # Run the job through the IBMQClient directly.
         api = backend._api
-        job = api.run_job(qobj.as_dict(), backend_name)
+        job = api.submit_job(qobj.as_dict(), backend_name)
 
         self.assertIn('status', job)
         self.assertIsNotNone(job['status'])
 
     @requires_qe_access
+    @requires_new_api_auth
+    def test_run_job_object_storage(self, qe_token, qe_url):
+        """Test running a job against a simulator using object storage."""
+        IBMQ.enable_account(qe_token, qe_url)
+
+        # Create a Qobj.
+        backend_name = 'ibmq_qasm_simulator'
+        backend = IBMQ.get_backend(backend_name)
+        circuit = transpile(self.qc1, backend, seed_transpiler=self.seed)
+        qobj = assemble(circuit, backend, shots=1)
+        print('complied')
+
+        # Run the job through the IBMQClient directly using object storage.
+        api = backend._api
+        job = api.job_submit_object_storage(backend_name, qobj.as_dict())
+        job_id = job['id']
+        self.assertEqual(job['kind'], 'q-object-external-storage')
+
+        # Wait for completion.
+        api.job_final_status_websocket(job_id)
+
+        # Fetch results and qobj via object storage.
+        result = api.job_result_object_storage(job_id)
+        qobj_downloaded = api.job_download_qobj_object_storage(job_id)
+
+        self.assertEqual(qobj_downloaded, qobj.as_dict())
+        self.assertEqual(result['status'], 'COMPLETED')
+
+    @requires_qe_access
+    @requires_new_api_auth
     def test_get_status_jobs(self, qe_token, qe_url):
         """Check get status jobs by user authenticated."""
         api = self._get_client(qe_token, qe_url)
@@ -83,6 +113,7 @@ class TestIBMQClient(QiskitTestCase):
         self.assertEqual(len(jobs), 2)
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_backend_status(self, qe_token, qe_url):
         """Check the status of a real chip."""
         backend_name = ('ibmq_20_tokyo'
@@ -92,6 +123,7 @@ class TestIBMQClient(QiskitTestCase):
         self.assertIsNotNone(is_available['operational'])
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_backend_properties(self, qe_token, qe_url):
         """Check the properties of calibration of a real chip."""
         backend_name = ('ibmq_20_tokyo'
@@ -102,6 +134,7 @@ class TestIBMQClient(QiskitTestCase):
         self.assertIsNotNone(properties)
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_available_backends(self, qe_token, qe_url):
         """Check the backends available."""
         api = self._get_client(qe_token, qe_url)
@@ -109,17 +142,18 @@ class TestIBMQClient(QiskitTestCase):
         self.assertGreaterEqual(len(backends), 1)
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_api_version(self, qe_token, qe_url):
         """Check the version of the QX API."""
         api = self._get_client(qe_token, qe_url)
         version = api.api_version()
         self.assertIsNotNone(version)
 
-    @skip('TODO: reenable after API check')
+    @skip('TODO: reenable after checking with API team')
     @requires_qe_access
+    @requires_new_api_auth
     def test_get_job_includes(self, qe_token, qe_url):
         """Check the field includes parameter for get_job."""
-        _ = self._get_client(qe_token, qe_url)
         IBMQ.enable_account(qe_token, qe_url)
 
         backend_name = 'ibmq_qasm_simulator'
@@ -128,7 +162,7 @@ class TestIBMQClient(QiskitTestCase):
         qobj = assemble(circuit, backend, shots=1)
 
         api = backend._api
-        job = api.run_job(qobj.as_dict(), backend_name)
+        job = api.submit_job(qobj.as_dict(), backend_name)
         job_id = job['id']
 
         # Get the job, excluding a parameter.
@@ -145,6 +179,7 @@ class TestAuthentication(QiskitTestCase):
     """
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_url_404(self, qe_token, qe_url):
         """Test login against a 404 URL"""
         url_404 = re.sub(r'/api.*$', '/api/TEST_404', qe_url)
@@ -152,6 +187,7 @@ class TestAuthentication(QiskitTestCase):
             _ = IBMQClient(qe_token, url_404)
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_invalid_token(self, qe_token, qe_url):
         """Test login using invalid token."""
         qe_token = 'INVALID_TOKEN'
@@ -159,6 +195,7 @@ class TestAuthentication(QiskitTestCase):
             _ = IBMQClient(qe_token, qe_url)
 
     @requires_qe_access
+    @requires_new_api_auth
     def test_url_unreachable(self, qe_token, qe_url):
         """Test login against an invalid (malformed) URL."""
         qe_url = 'INVALID_URL'